requirements.txt
python-telegram-bot==20.6
pydantic==2.3.0
motor==3.3.0
pymongo==4.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
loguru==0.7.2 